requirements.txt
Augmentor==0.2.10
imageio==2.22.4
ipython==8.7.0
keras==2.9.0
matplotlib==3.5.2
numpy==1.23.1
opencv_python==4.6.0.66
pandas==1.5.2
Pillow==9.3.0
scikit_learn==1.1.3
scipy==1.9.1
seaborn==0.12.1
tensorflow==2.9.1
